réseaux ressources : - 4819721 action magique changer capacités des liens - mantis 4027: les cellules des tables ne sont pas automatiquement validées quand on clique sur valider - cleaning and minor fixes

+++ src/libraries/gaming/UrbanKnowledge.h
@@ -91,8 +91,8 @@ private:
     urban::TerrainObject_ABC* pRealUrban_;
 
     kernel::OptionalValue< bool > bIsPerceived_;
-    kernel::OptionalValue< float > rProgress_;
-    kernel::OptionalValue< float > rMaxProgress_;
+    kernel::OptionalValue< int > rProgress_;
+    kernel::OptionalValue< int > rMaxProgress_;
     //@}
 };
